+++ crates/cubecl-linalg/src/convolution/homogeneous/multi_stage_tma.rs
@@ -181,7 +181,7 @@ where
 
         sync_cube();
 
-        SMM::read_accumulator::<Self::Config>(acc, &mut out_writer, config.to_smm_config(), config);
+        SMM::write_results::<Self::Config>(acc, &mut out_writer, config.to_smm_config(), config);
     }
 
     fn init_lhs_loader(
