+++ crates/cubecl-linalg/src/matmul/components/global/copy_mechanism.rs
@@ -3,9 +3,7 @@ use cubecl_core::prelude::barrier::Barrier;
 use cubecl_core::prelude::*;
 
 #[cube]
-// TODO remove copy clone
-// pub trait CopyMechanism<ES: Numeric>: CubeType + Sync + Send + 'static {
-pub trait CopyMechanism<ES: Numeric>: CubeType + Clone + Copy + Sync + Send + 'static {
+pub trait CopyMechanism<ES: Numeric>: CubeType + Sync + Send + 'static {
     fn memcpy_async(this: &Self, source: &Slice<Line<ES>>, destination: &mut SliceMut<Line<ES>>);
 }
