+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/double_buffering/matmul.rs
@@ -223,7 +223,7 @@ where
             config.to_smm_config(),
         );
 
-        SMM::read_accumulator::<Self::Config>(acc, &mut out_writer, config.to_smm_config(), config);
+        SMM::write_results::<Self::Config>(acc, &mut out_writer, config.to_smm_config(), config);
     }
 
     fn init_lhs_loader(
