+++ crates/cubecl-linalg/src/matmul/components/global/write/mod.rs
@@ -1,5 +1,7 @@
+mod base;
 mod tilewise;
 mod unit;
 
+pub use base::*;
 pub use tilewise::TilewiseWriter;
 pub use unit::UnitWriter;
