+++ crates/cubecl-linalg/src/matmul/components/global/write/unit.rs
@@ -1,10 +1,11 @@
 use crate::matmul::components::Ident;
 use crate::matmul::components::global::GlobalConfig;
 use crate::matmul::components::global::tensor_view::TensorWriter;
-use crate::matmul::components::stage::GlobalWriter;
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 use cubecl_std::tensor::r#virtual::{ReadWrite, VirtualTensor};
+
+use super::GlobalWriter;
 
 #[derive(CubeType)]
 pub struct UnitWriter<EG: Numeric> {
